pyproject.toml
[tool.black]
line-length = 100
target-version = ['py37']

[tool.isort]
known_third_party = [
  "black",
  "click",
  "click_option_group",
  "hydra",
  "matplotlib",
  "mypy",
  "numpy",
  "omegaconf",
  "pandas",
  "pylint",
  "pytest",
  "scipy",
  "setuptools",
  "sklearn",
  "torch",
  "tqdm",
  "wandb",
]
known_future_library = []
extra_standard_library = ["dataclasses", "__future__"]
line_length = 88
profile = "black"
force_sort_within_sections = "True"
classes = ["MISSING"]

[tool.poetry]
name = "whaledo"
version = "0.0.1"
description = ""
authors = ["Finding Bailey"]
license = "Apache License 2.0"
packages = [
  { include = "src" },
  { include = "hydra_plugins" },
]
classifiers=[
  "Programming Language :: Python :: 3.9",
  "Programming Language :: Python :: 3.10",
  "License :: OSI Approved :: Apache Software License",
  "Operating System :: OS Independent",
  "Typing :: Typed",
]
keywords=["typing", "python"]
repository=""
readme="README.md"

[tool.poetry.dependencies]
python = "3.9.7"
torch = "^1.11"
ranzen = {version = "^1.2.6", extras = ["all"]}
scikit-learn = "^0.24.1"
wandb = "^0.12.10"
attrs = "^21.4.0"
faiss-cpu = "^1.7.0"
classy-vision = "^0.6.0"
setuptools = "59.5.0" # for pytorch-lightning
fairscale = "^0.4.5"
hydra-submitit-launcher = "^1.1.6"
typing-extensions = ">=4.0"
deepspeed = { version = "^0.5.10", platform = "linux", optional = true }
triton = { version = "1.0.0", platform = "linux",  optional = true }
wrapt = "<1.14,>=1.11"
timm = "^0.5.4"
omegaconf = "2.1.2"
einops = "^0.4.1"
augly = "^1.0.0"
albumentations = "^1.1.0"
transformers = "^4.19.2"
torch-conduit = "^0.1.1"

[tool.poetry.extras]
deepspeed = ["deepspeed", "triton"]

[tool.poetry.dev-dependencies]
black = "^22.1"
pytest = "^6.2.2"
isort = "^5.7.0"
darglint = "^1.8.0"

[tool.pyright]
exclude = ["**/node_modules", "**/__pycache__", "**/.**"]
typeCheckingMode = "basic"
pythonVersion = "3.10"
reportUnusedImport = "error"
reportDuplicateImport  = "error"
reportIncompatibleVariableOverride = "error"
reportOverlappingOverload = "error"
reportUntypedNamedTuple = "error"
reportConstantRedefinition = "warning"
reportMissingImports = "error"
reportMissingTypeStubs = "warning"
strictListInference = true
strictSetInference = true
strictParameterNoneValue = true

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
